+++ docs/pages/components/pagination/index.tsx
@@ -1,7 +1,31 @@
 import React from 'react';
-import { Pagination, Button, Toggle, Divider } from 'rsuite';
+import {
+  Pagination,
+  Button,
+  Toggle,
+  Divider,
+  Slider,
+  SelectPicker,
+  TagPicker,
+  InputNumber
+} from 'rsuite';
+import Link from 'next/link';
 import DefaultPage from '@/components/Page';
 
 export default function Page() {
-  return <DefaultPage dependencies={{ Pagination, Button, Toggle, Divider }} />;
+  return (
+    <DefaultPage
+      dependencies={{
+        Link,
+        Pagination,
+        Button,
+        Toggle,
+        Divider,
+        Slider,
+        SelectPicker,
+        TagPicker,
+        InputNumber
+      }}
+    />
+  );
 }